feat: add report feature and backend adjustments

+++ backend/wall-common/src/main/java/cn/yiming1234/NottinghamWall/context/BaseContext.java
@@ -2,14 +2,14 @@ package cn.yiming1234.NottinghamWall.context;
 
 public class BaseContext {
 
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
-        return threadLocal.get();
+    public static Integer getCurrentId() {
+        return Math.toIntExact(threadLocal.get());
     }
 
     public static void removeCurrentId() {
